Some more .java files have been added. Validation for existing tables in the database has also been added.
Global lock of the type of data that the files will be working with
still needs to be worked on.

Problems with FileIO of the .csv files. Might just use Java methods
for loading .csv files straight too Database. Will need to discuss this
approach. Once the FileIO is done. Development on the queries can begin.